feat: simplify project by replacing ItemWrapper with CarouselCard and integrating dialog
